Extract generic decorator into a reusable class
The only current use of the decorator pattern in colcon-core is the
argument parser, but the concept can be applied generically in a variety
of scenarios.
